+++ frontend/src/pages/Home.jsx
@@ -62,7 +62,6 @@ export default function Home() {
         <div className="product-list">
           {productsToDisplayList?.map((product) => (
             <ProductListCard key={product.id} product={product} />
-              
           ))}
         </div>
       </section>
