+++ frontend/src/pages/profile/AccountCreated.jsx
@@ -1,5 +1,5 @@
 import { UserCircle } from "lucide-react";
-import PageIconCircle from "../components/PageIconCircle";
+import PageIconCircle from "../../components/PageIconCircle";
 import { Link } from "react-router-dom";
 
 function AccountCreated() {
