+++ frontend/src/pages/profile/ResetPassword.jsx
@@ -1,6 +1,6 @@
 import { EyeOff } from "lucide-react";
 import { Link } from "react-router-dom";
-import ThemedInput from "../components/ThemedInput";
+import ThemedInput from "../../components/ThemedInput";
 
 function ResetPassword() {
   return (
