+++ frontend/src/pages/profile/SignIn.jsx
@@ -1,6 +1,6 @@
 import { Check, EyeOff, Facebook, Twitter } from "lucide-react";
 import { Link } from "react-router-dom";
-import ThemedInput from "../components/ThemedInput";
+import ThemedInput from "../../components/ThemedInput";
 
 export default function SignIn() {
   return (
