+++ frontend/src/routes.jsx
@@ -1,20 +1,20 @@
-import NotFound from "./pages/NotFound";
-import Cart from "./pages/Cart";
-import SignIn from "./pages/SignIn";
 import Home from "./pages/Home";
-import Order from "./pages/Checkout";
-import ProductDetails from "./pages/ProductDetails";
-import Search from "./pages/Search";
-import Products from "./pages/Products";
-import OrderConfirmation from "./pages/OrderConfirmation";
-import ShippingDetails from "./pages/ShippingDetails";
-import AddPaymentCard from "./pages/AddPaymentCard";
-import WishList from "./pages/WishList";
-import ForgotPassword from "./pages/ForgotPassword";
-import ResetPassword from "./pages/ResetPassword";
-import PasswordConfirmation from "./pages/PasswordConfirmation";
-import SignUp from "./pages/SignUp";
-import AccountCreated from "./pages/AccountCreated";
+import Search from "./pages/products/Search";
+import SignIn from "./pages/profile/SignIn";
+import SignUp from "./pages/profile/SignUp";
+import ForgotPassword from "./pages/profile/ForgotPassword";
+import ResetPassword from "./pages/profile/ResetPassword";
+import PasswordConfirmation from "./pages/profile/PasswordConfirmation";
+import AccountCreated from "./pages/profile/AccountCreated";
+import Cart from "./pages/order/Cart";
+import Products from "./pages/products/Products";
+import ProductDetails from "./pages/products/ProductDetails";
+import ShippingDetails from "./pages/order/ShippingDetails";
+import WishList from "./pages/products/WishList";
+import OrderConfirmation from "./pages/order/OrderConfirmation";
+import AddPaymentCard from "./pages/order/AddPaymentCard";
+import NotFound from "./pages/NotFound";
+import Checkout from "./pages/order/Checkout";
 
 // If no title is set, the manero logo will show instead
 const routes = [
@@ -29,7 +29,7 @@ const routes = [
   { path: "/cart", element: <Cart /> },
   { path: "/products", element: <Products />, title: "Products" },
   { path: "/products/:id", element: <ProductDetails /> },
-  { path: "/checkout", element: <Order />, title: "Checkout" },
+  { path: "/checkout", element: <Checkout />, title: "Checkout" },
   { path: "/checkout/shipping", element: <ShippingDetails />, title: "Shipping" },
   { path: "/wishlist", element: <WishList />, title: "Wishlist" },
   {
