+++ frontend/src/scss/main.scss
@@ -24,16 +24,16 @@
 // Pages
 @use "pages/not-found";
 @use "pages/home";
-@use "pages/search";
-@use "pages/cart";
-@use "pages/product-details";
-@use "pages/products";
-@use "pages/checkout";
-@use "pages/order-confirmation";
-@use "pages/shipping-details";
-@use "pages/add-payment-card";
-@use "pages/sign-in";
-@use "pages/sign-up";
-@use "pages/wishlist";
-@use "pages/forgot-password";
-@use "pages/reset-password";
+@use "pages/products/search";
+@use "pages/products/product-details";
+@use "pages/products/products";
+@use "pages/products/wishlist";
+@use "pages/order/checkout";
+@use "pages/order/cart";
+@use "pages/order/order-confirmation";
+@use "pages/order/shipping-details";
+@use "pages/order/add-payment-card";
+@use "pages/profile/sign-in";
+@use "pages/profile/sign-up";
+@use "pages/profile/forgot-password";
+@use "pages/profile/reset-password";
