+++ frontend/src/scss/pages/order/_add-payment-card.scss
@@ -1,4 +1,4 @@
-@use "../core/breakpoints" as *;
+@use "../../core/breakpoints" as *;
 
 .add-payment-card-page {
   padding-inline: var(--section-padding-inline);
