+++ frontend/src/scss/pages/order/_cart.scss
@@ -1,4 +1,4 @@
-@use "../core/breakpoints" as *;
+@use "../../core/breakpoints" as *;
 
 .cart-page {
   padding-block: 2rem;
