+++ frontend/src/scss/pages/order/_checkout.scss
@@ -1,4 +1,4 @@
-@use "../core/breakpoints" as *;
+@use "../../core/breakpoints" as *;
 
 .checkout-page {
   h2 {
