+++ frontend/src/scss/pages/order/_order-confirmation.scss
@@ -1,4 +1,4 @@
-@use "../core/breakpoints" as *;
+@use "../../core/breakpoints" as *;
 
 .order-confirmation-page {
   padding-block: 2rem;
