+++ frontend/src/scss/pages/order/_shipping-details.scss
@@ -1,4 +1,4 @@
-@use "../core/breakpoints" as *;
+@use "../../core/breakpoints" as *;
 
 .shipping-details-page {
   padding-inline: var(--section-padding-inline);
