+++ frontend/src/scss/pages/products/_product-details.scss
@@ -1,4 +1,4 @@
-@use "../core/breakpoints" as *;
+@use "../../core/breakpoints" as *;
 
 .product-details-page {
   @include md {
