+++ frontend/src/scss/pages/products/_products.scss
@@ -1,4 +1,4 @@
-@use "../core/breakpoints" as *;
+@use "../../core/breakpoints" as *;
 
 .products-page {
   padding-inline: var(--section-padding-inline);
