+++ frontend/src/scss/pages/products/_search.scss
@@ -1,4 +1,4 @@
-@use "../core/breakpoints" as *;
+@use "../../core/breakpoints" as *;
 
 .search-page {
   .tags-container {
