+++ frontend/src/scss/pages/profile/_forgot-password.scss
