+++ frontend/src/scss/pages/profile/_reset-password.scss
