+++ frontend/src/scss/pages/profile/_sign-up.scss
